* Approx. 18 % faster index building for AVCHD streams (and others)
git-svn-id: svn://svn.code.sf.net/p/gmerlin/code/trunk@2363 214d8ac5-8513-4f59-80f8-2f39b674a24a

+++ gmerlin_avdecoder/include/bitstream.h
@@ -26,7 +26,7 @@ typedef struct
   const uint8_t * pos;
   const uint8_t * end;
   int bit_cache;
-  uint8_t c;
+  uint32_t c;
   } bgav_bitstream_t;
 
 void bgav_bitstream_init(bgav_bitstream_t * b, const uint8_t * pos,
